refactor: use viem for chains
