+++ packages/frontend/src/providers/SelectedChainProvider.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useState, useContext } from 'react';
+import { SUPPORTED_CHAINS } from 'src/utils/chains';
 import { Chain } from 'wagmi';
-import { SUPPORTED_CHAINS } from '../utils/evm';
 
 type SelectedChainContextType = {
   selectedChain: Chain & { icon: string };
@@ -10,7 +10,7 @@ type SelectedChainContextType = {
 const SelectedChainContext = createContext<SelectedChainContextType | undefined>(undefined);
 
 const SelectedChainProvider: React.FC<React.PropsWithChildren> = ({ children }) => {
-  const [selectedChain, setSelectedChain] = useState(SUPPORTED_CHAINS.ethereum);
+  const [selectedChain, setSelectedChain] = useState(SUPPORTED_CHAINS.mainnet);
 
   return (
     <SelectedChainContext.Provider value={{ selectedChain, setSelectedChain }}>
